Always mark the last variant as default
CLDR plurals and StringBundle's plurals are not compatible because CLDR
accounts for fractions.  In some languages (e.g. Polish, go figure) 'many' is
for large integers and 'other' is for fractions.  It means that for some
languages CLDR has one category more than StringBundle.

For instance, CLDR returns ('one', 'few', 'many', 'other') for Polish while
the Polish localization uses the rule #9 with only three forms.  We'll never
reach 'other' if we iterate over the legacy variants.

This patch changes the code to always set default on the last variant rather
than the one whose key is 'other'.