Refactor the extended AST.
This is a big refactor of how the transform AST works.  Previously the
transforms were evaluated instantly at the moment when they were added to the
context.  This required helpers like MESSAGE and SOURCE to be partials bound to
the state of the already existing context.

    MESSAGE = ctx.create_message()
    SOURCE = ctx.create_source()

    ctx.add_transforms([
        MESSAGE('aboutDownloads.ftl', 'title')(
            value=COPY(
                SOURCE(
                   'aboutDownloads.dtd',
                    'aboutDownloads.title'
                )
            )
        )
    ]

After the refactor, the extended AST is just a raw and lazy representation of
the transforms required to migrate a message.  When a transform is added, its
ASt is scanned for any SOURCE nodes to build a dependency graph.

    ctx.add_transforms('aboutDownloads.ftl', [
        FTL.Entity(
            id=FTL.Identifier('title'),
            value=COPY(
                SOURCE(
                   'aboutDownloads.dtd',
                    'aboutDownloads.title'
                )
            )
        )
    ]

A transform are evaluated explicitly only when a migration is run using
`MergeContext.serialize_changeset`.